+++ src/org/jitsi/impl/neomedia/transform/rewriting/ExtendedSequenceNumberInterval.java
@@ -92,8 +92,8 @@ class ExtendedSequenceNumberInterval
     {
         public boolean test(REDBlock redBlock)
         {
-            Map<Integer, Byte> ssrc2fec = getSsrcRewritingEngine().ssrc2fec;
-            int sourceSSRC = ssrcRewriter.getSourceSSRC();
+            Map<Long, Byte> ssrc2fec = getSsrcRewritingEngine().ssrc2fec;
+            long sourceSSRC = ssrcRewriter.getSourceSSRC();
             return redBlock != null
                 && ssrc2fec.get(sourceSSRC) == redBlock.getPayloadType();
         }
@@ -164,9 +164,9 @@ class ExtendedSequenceNumberInterval
     {
         // SSRC
         SsrcGroupRewriter ssrcGroupRewriter = getSsrcGroupRewriter();
-        int ssrcTarget = ssrcGroupRewriter.getSSRCTarget();
+        long ssrcTarget = ssrcGroupRewriter.getSSRCTarget();
 
-        pkt.setSSRC(ssrcTarget);
+        pkt.setSSRC((int) ssrcTarget);
 
         // Sequence number
         int seqnum = pkt.getSequenceNumber();
@@ -192,9 +192,9 @@ class ExtendedSequenceNumberInterval
 
         SsrcRewritingEngine ssrcRewritingEngine
             = ssrcGroupRewriter.ssrcRewritingEngine;
-        Map<Integer, Integer> rtx2primary = ssrcRewritingEngine.rtx2primary;
-        int sourceSSRC = ssrcRewriter.getSourceSSRC();
-        Integer primarySSRC = rtx2primary.get(sourceSSRC);
+        Map<Long, Long> rtx2primary = ssrcRewritingEngine.rtx2primary;
+        long sourceSSRC = ssrcRewriter.getSourceSSRC();
+        Long primarySSRC = rtx2primary.get(sourceSSRC);
 
         if (primarySSRC == null)
             primarySSRC = sourceSSRC;
@@ -247,8 +247,8 @@ class ExtendedSequenceNumberInterval
     {
         // This is an RTX packet. Replace RTX OSN field or drop.
         SsrcRewritingEngine ssrcRewritingEngine = getSsrcRewritingEngine();
-        int sourceSSRC = ssrcRewriter.getSourceSSRC();
-        int ssrcOrig = ssrcRewritingEngine.rtx2primary.get(sourceSSRC);
+        long sourceSSRC = ssrcRewriter.getSourceSSRC();
+        long ssrcOrig = ssrcRewritingEngine.rtx2primary.get(sourceSSRC);
         int snOrig = pkt.getOriginalSequenceNumber();
 
         SsrcGroupRewriter rewriterPrimary
@@ -289,7 +289,7 @@ class ExtendedSequenceNumberInterval
      * @return {@code true} if the RED was successfully rewritten;
      * {@code false}, otherwise
      */
-    private boolean rewriteRED(int primarySSRC, byte[] buf, int off, int len)
+    private boolean rewriteRED(long primarySSRC, byte[] buf, int off, int len)
     {
         if (buf == null || buf.length == 0)
         {
@@ -328,7 +328,7 @@ class ExtendedSequenceNumberInterval
      * @return {@code true} if the FEC was successfully rewritten;
      * {@code false}, otherwise
      */
-    private boolean rewriteFEC(int sourceSSRC, byte[] buf, int off, int len)
+    private boolean rewriteFEC(long sourceSSRC, byte[] buf, int off, int len)
     {
         if (buf == null || buf.length == 0)
         {
